+++ R/utils.R
@@ -1,3 +1,40 @@
+#' Assert whether object belongs to a class
+#'
+#' @description Checks if an object belongs to a certain class or classes. If
+#' not, stops with an error message.
+#' @param x object to be checked
+#' @param y character; class name(s)
+#' @references Copied from \url{https://github.com/ropensci/taxizedb}
+#' @noRd
+#' @examples
+#' assert(1:10, "integer")
+#' assert(data.frame(a = 1:10), c("data.frame", "list"))
+assert <- function(x, y) {
+  if (!is.null(x)) {
+    if (!inherits(x, y)) {
+      stop(deparse(substitute(x)), " must be of class ",
+           paste0(y, collapse = ", "), call. = FALSE)
+    }
+  }
+}
+
+#' Get URLs and file names of local database files
+#'
+#' @param db character; database name. Currently only "chembl" is supported.
+#' @param version character; version of the database. Either "latest" (default)
+#' or a specific version number, e.g. "30".
+#' @return A data frame. Variables depend on the database.
+#' @examples
+#' db_files("chembl", version = "latest")
+#' db_files("chembl", version = "30")
+#' @noRd
+db_files <- function(db, version = "latest") {
+  db <- match.arg(db, choices = c("chembl"))
+  if (db == "chembl") {
+    chembl_files(version = version)
+  }
+}
+
 #' Check if input is a valid inchikey
 #'
 #' @description This function checks if a string is a valid inchikey.
@@ -536,6 +573,20 @@ matcher <-
       }
     }
   }
+
+#' Check if an url exists
+#'
+#' @param url url
+#' @noRd
+url_exists <- function(url) {
+  foo <- function(x) {
+    res <- try(httr::HEAD(x), silent = TRUE)
+    if (inherits(res, "try-error")) return(FALSE)
+    status <- httr::status_code(res)
+    status >= 200 && status < 400
+  }
+  unname(sapply(url, foo))
+}
 
 #' Webchem messages
 #'